refactor: 형광펜 관련 api 파일명 변경(collection -> highlight)
