+++ frontend/src/mocks/handlers/index.ts
@@ -1,7 +1,7 @@
-import collectionHandler from './collection';
 import groupHandler from './group';
+import highlightHandler from './highlight';
 import reviewHandler from './review';
 
-const handlers = [...reviewHandler, ...groupHandler, ...collectionHandler];
+const handlers = [...reviewHandler, ...groupHandler, ...highlightHandler];
 
 export default handlers;
